\name{checkData}
\alias{checkData}
\title{Perform a data check}
\description{
Performs various data checks to see if the data are applicable for Guttman error tree analysis.
}
\usage{
checkData(data)
}
\arguments{
  \item{data}{
  a data.frame or matrix.
}
}
\details{
Function was adapted from the mokken Package (van der Ark, 2007).
}
\value{
Returns a data.frame or throws an error if something is wrong with the dataset.
}
\references{
Van der Ark, L. A. (2007). Mokken scale analysis in R. Journal of Statistical Software. http://www.jstatsoft.org/v20/i11
}

\examples{
###
1+1
}
