\name{Communality}
\alias{Communality}
\docType{data}
\title{Communality Data}

\description{
Dataset containing responses to one of the Adjective Checklist scales.
}
\usage{data(Communality)}

\source{
 Data were adapted from the mokken Package (van der Ark, 2007).
}
\references{
Van der Ark, L. A. (2007) 
Mokken scale analysis in R. 
\emph{Journal of Statistical Software}.
\url{http://www.jstatsoft.org/v20/i11}
}

\examples{
data(Communality)
}
