\name{GetR-package}
\alias{GetR-package}
\alias{GetR}
\docType{package}
\title{
GetR: Guttman error trees in R
}
\description{
The GetR package calculates Guttman error trees, which can be used to find homogeneous subgroups regarding Guttman errors.
}
\details{
\tabular{ll}{
Package: \tab GetR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-02-16\cr
License: \tab GPL-3\cr
}
}
\author{
Johannes Beller, Soeren Kliem

Maintainer: Johannes Beller <johannesbeller@gmail.com>
}

\keyword{ package }